common parameters in gguidance

This is a collection of common parameters so they needn't be re-documented each time.

Arguments

title

A <character[1]> or <expression[1]> indicating the title of the guide. If NULL, the title is not shown. The default, waiver(), takes the name of the scale object or the name specified in labs() as the title.

theme

A <theme> object to style the guide individually or differently from the plot's theme settings. The theme argument in the guide overrides and is combined with the plot's theme.

position

A <character[1]> giving the location of the guide. Can be one of "top", "bottom", "left" or "right".

order

A positive <integer[1]> that specifies the order of this guide among multiple guides. This controls in which order guides are merged if there are multiple guides for the same position. If 0 (default), the order is determined by a hashing indicative settings of a guide.

available_aes

A <character> vector listing the aesthetics for which this guide can be build.

direction

A <character[1]> indicating the direction of the guide. Can be on of "horizontal" or "vertical".

angle

A specification for the text angle. Compared to setting the angle argument in element_text(), this argument uses some heuristics to automatically pick the hjust and vjust that you probably want. Can be one of the following:

NULL to take angles and justification settings directly from the theme.

waiver() to allow reasonable defaults in special cases.

A <numeric[1]> between -360 and 360 for the text angle in degrees.