Guide composition

Guide composition is a meta-guide orchestrating an ensemble of other guides. On their own, a 'composing' guide is not very useful as a visual reflection of a scale.

Usage

new_compose(
  guides,
  args = list(),
  ...,
  available_aes = c("any", "x", "y", "r", "theta"),
  call = caller_env(),
  super = Compose
)

Arguments

guides

A <list> of guides wherein each element is one of the following:

A <Guide> class object.

A <function> that returns a <Guide> class object.

A <character[1]> naming such a function, without the guide_ or primitive_ prefix.

args

A <list> of arguments to pass to guides that are given either as a function or as a string.

...

Additional parameters to pass on to new_guide().

available_aes

A <character> giving aesthetics that must match the the guides.

call

A call to display in messages.

super

A <Compose> class object giving a meta-guide for composition.

Value

A <Compose> (sub-)class guide that composes other guides.

See also

Examples

# new_compose() is not intended to be used directly